requirements.txt
streamlit==1.45.1
opencv-python-headless==4.11.0.86
numpy==2.2.5
pillow==11.2.1
